+++ android/src/main/java/com/lib/flutter_blue_plus/FlutterBluePlusPlugin.java
@@ -2212,7 +2212,7 @@ public class FlutterBluePlusPlugin implements
         @TargetApi(33) // newer function with byte[] value argument
         public void onCharacteristicRead(@NonNull BluetoothGatt gatt, BluetoothGattCharacteristic characteristic, @NonNull byte[] value, int status) {
             // this callback is only for explicit characteristic reads
-            LogLevel level = status == 0 ? LogLevel.DEBUG : LogLevel.ERROR;
+            LogLevel level = status == BluetoothGatt.GATT_SUCCESS ? LogLevel.DEBUG : LogLevel.ERROR;
             log(level, "onCharacteristicRead:");
             log(level, "  chr: " + uuidStr(characteristic.getUuid()));
             log(level, "  status: " + gattErrorString(status) + " (" + status + ")");
@@ -2220,8 +2220,9 @@ public class FlutterBluePlusPlugin implements
         }
 
         @Override
-        public void onCharacteristicWrite(BluetoothGatt gatt, BluetoothGattCharacteristic characteristic, int status) {
-            LogLevel level = status == 0 ? LogLevel.DEBUG : LogLevel.ERROR;
+        public void onCharacteristicWrite(BluetoothGatt gatt, BluetoothGattCharacteristic characteristic, int status)
+        {
+            LogLevel level = status == BluetoothGatt.GATT_SUCCESS ? LogLevel.DEBUG : LogLevel.ERROR;
             log(level, "onCharacteristicWrite:");
             log(level, "  chr: " + uuidStr(characteristic.getUuid()));
             log(level, "  status: " + gattErrorString(status) + " (" + status + ")");
@@ -2794,49 +2795,43 @@ public class FlutterBluePlusPlugin implements
 
     // Defined in the Bluetooth Standard
     private static String gattErrorString(int value) {
-        switch (value) {
-            case BluetoothGatt.GATT_SUCCESS:
-                return "GATT_SUCCESS";                     // 0
-            case 0x01:
-                return "GATT_INVALID_HANDLE";              // 1
-            case BluetoothGatt.GATT_READ_NOT_PERMITTED:
-                return "GATT_READ_NOT_PERMITTED";          // 2
-            case BluetoothGatt.GATT_WRITE_NOT_PERMITTED:
-                return "GATT_WRITE_NOT_PERMITTED";         // 3
-            case 0x04:
-                return "GATT_INVALID_PDU";                 // 4
-            case BluetoothGatt.GATT_INSUFFICIENT_AUTHENTICATION:
-                return "GATT_INSUFFICIENT_AUTHENTICATION"; // 5
-            case BluetoothGatt.GATT_REQUEST_NOT_SUPPORTED:
-                return "GATT_REQUEST_NOT_SUPPORTED";       // 6
-            case BluetoothGatt.GATT_INVALID_OFFSET:
-                return "GATT_INVALID_OFFSET";              // 7
-            case BluetoothGatt.GATT_INSUFFICIENT_AUTHORIZATION:
-                return "GATT_INSUFFICIENT_AUTHORIZATION";  // 8
-            case 0x09:
-                return "GATT_PREPARE_QUEUE_FULL";          // 9
-            case 0x0a:
-                return "GATT_ATTR_NOT_FOUND";              // 10
-            case 0x0b:
-                return "GATT_ATTR_NOT_LONG";               // 11
-            case 0x0c:
-                return "GATT_INSUFFICIENT_KEY_SIZE";       // 12
-            case BluetoothGatt.GATT_INVALID_ATTRIBUTE_LENGTH:
-                return "GATT_INVALID_ATTRIBUTE_LENGTH";    // 13
-            case 0x0e:
-                return "GATT_UNLIKELY";                    // 14
-            case BluetoothGatt.GATT_INSUFFICIENT_ENCRYPTION:
-                return "GATT_INSUFFICIENT_ENCRYPTION";     // 15
-            case 0x10:
-                return "GATT_UNSUPPORTED_GROUP";           // 16
-            case 0x11:
-                return "GATT_INSUFFICIENT_RESOURCES";      // 17
-            case BluetoothGatt.GATT_CONNECTION_CONGESTED:
-                return "GATT_CONNECTION_CONGESTED";        // 143
-            case BluetoothGatt.GATT_FAILURE:
-                return "GATT_FAILURE";                     // 257
-            default:
-                return "UNKNOWN_GATT_ERROR (" + value + ")";
+        switch(value) {
+            case BluetoothGatt.GATT_SUCCESS                     : return "GATT_SUCCESS";                     // 0
+            case 0x01                                           : return "GATT_INVALID_HANDLE";              // 1
+            case BluetoothGatt.GATT_READ_NOT_PERMITTED          : return "GATT_READ_NOT_PERMITTED";          // 2
+            case BluetoothGatt.GATT_WRITE_NOT_PERMITTED         : return "GATT_WRITE_NOT_PERMITTED";         // 3
+            case 0x04                                           : return "GATT_INVALID_PDU";                 // 4
+            case BluetoothGatt.GATT_INSUFFICIENT_AUTHENTICATION : return "GATT_INSUFFICIENT_AUTHENTICATION"; // 5
+            case BluetoothGatt.GATT_REQUEST_NOT_SUPPORTED       : return "GATT_REQUEST_NOT_SUPPORTED";       // 6
+            case BluetoothGatt.GATT_INVALID_OFFSET              : return "GATT_INVALID_OFFSET";              // 7
+            case BluetoothGatt.GATT_INSUFFICIENT_AUTHORIZATION  : return "GATT_INSUFFICIENT_AUTHORIZATION";  // 8
+            case 0x09                                           : return "GATT_PREPARE_QUEUE_FULL";          // 9
+            case 0x0a                                           : return "GATT_ATTR_NOT_FOUND";              // 10
+            case 0x0b                                           : return "GATT_ATTR_NOT_LONG";               // 11
+            case 0x0c                                           : return "GATT_INSUFFICIENT_KEY_SIZE";       // 12
+            case BluetoothGatt.GATT_INVALID_ATTRIBUTE_LENGTH    : return "GATT_INVALID_ATTRIBUTE_LENGTH";    // 13
+            case 0x0e                                           : return "GATT_UNLIKELY";                    // 14
+            case BluetoothGatt.GATT_INSUFFICIENT_ENCRYPTION     : return "GATT_INSUFFICIENT_ENCRYPTION";     // 15
+            case 0x10                                           : return "GATT_UNSUPPORTED_GROUP";           // 16
+            case 0x11                                           : return "GATT_INSUFFICIENT_RESOURCES";      // 17
+            case 0x80                                           : return "GATT_NO_RESOURCES";                // 128
+            case 0x81                                           : return "GATT_INTERNAL_ERROR";              // 129
+            case 0x82                                           : return "GATT_WRONG_STATE";                 // 130
+            case 0x83                                           : return "GATT_DB_FULL";                     // 131
+            case 0x84                                           : return "GATT_BUSY";                        // 132
+            case 0x85                                           : return "GATT_ERROR";                       // 133
+            case 0x86                                           : return "GATT_CMD_STARTED";                 // 134
+            case 0x87                                           : return "GATT_ILLEGAL_PARAMETER";           // 135
+            case 0x88                                           : return "GATT_PENDING";                     // 136
+            case 0x89                                           : return "GATT_AUTH_FAIL";                   // 137
+            case 0x8a                                           : return "GATT_MORE";                        // 138
+            case 0x8b                                           : return "GATT_INVALID_CFG";                 // 139
+            case 0x8c                                           : return "GATT_SERVICE_STARTED";             // 140
+            case 0x8d                                           : return "GATT_ENCRYPTED_NO_MITM";           // 141
+            case 0x8e                                           : return "GATT_NOT_ENCRYPTED";               // 142
+            case BluetoothGatt.GATT_CONNECTION_CONGESTED        : return "GATT_CONNECTION_CONGESTED";        // 143
+            case BluetoothGatt.GATT_FAILURE                     : return "GATT_FAILURE";                     // 257
+            default: return "UNKNOWN_GATT_ERROR (" + value + ")";
         }
     }